magit-submodule-add: allow any path, read url before path
The old implementation was optimized for an odd special-case and did
not work in the normal case because it was only possible to select an
existing path.

Read the url before the path because the former is the only mandatory
argument.

Force the use of `read-file-name-default' because `ido-read-file-name'
behaves strangely (the users choice is ignored the first time and s/he
is prompted again, this time with a different initial input).