allow magit-repository-directories to be directory names
Adjust downstream processing of magit-repository-directories to let an
element of magit-repository-directories be a directory name (that is,
end in a slash) rather than a directory file name.  This prevents an
alist with empty strings for all keys from being passed to
magit-list-repos-uniquify, in which case the function believes that all
repos should be uniquified.