magit-section-goto-successor: use magit-section-goto for sibling hunks
When going to a sibling hunk of the previously current hunk which just
disappeared, then use `magit-section-goto' to go there.

In this case we actually did move to another section and we also want
to run `magit-section-movement-hook' for `magit-hunk-set-window-start'.

Other hook functions should not cause an issues as they either have no
affect when going to a hunk section.

Closes #2683.