magit-branch-or-commit-at-point: also consider tag at point
A tag is just a name for a revision, so everywhere a revision can be
used a tag should also be fine.

The worst that can happen is that for some callers it would feel more
natural to use a hash than a tag, but after briefly inspecting all the
existing callers inside Magit, I don't think this is ever actually the
case.

On the other hand not returning the tag at point represented by a
section of type `tag' (as opposed to a section of type `commit' which
happens to be tagged) often meant that commands which are supposed to
default to the branch or commit at point instead defaulted to some
fallback default.

Closes #2682.