magit-blame: ensure git process runs at repository top-level
Changing the buffer changes the value of `default-directory' even though
it is let-bound, because for buffer-local variables such bindings are
buffer-local too.

In [1] we ensured that the desired binding stayed in effect for indirect
buffers, now we extend this to blog buffers.  To make it more obvious
what is going on explicitly let-bind `default-directory' twice, but
avoid calculating the value twice.

Fixes #2731.