magit-log-output-coding-system: new option
This option serves the same purpose as
`vc-git-log-output-coding-system', which was added in Emacs v25.1.

The default value is always `utf-8', the global and local values of
`i18n.logOutputEncoding' are being ignored.  The local values are being
ignored because checking that all the time would be to costly.  We could
write a cache though.  The global value is being ignored because we
would first need a translation table, to turn e.g. Git's "EUCJP" into
Emacs' `euc-jp'.

Fixes #2738.