Support duet and other alternative commit commands
Git duet is a tool to support a pair-programming workflow with git:

https://github.com/git-duet/git-duet

It provides a handy interface for setting one of the pair as the commit
author, and the other of the pair as the committer. In this way both
members of the pair are attributed on github, and so on. Using git-duet
requires committing not with the command `git commit`, but with the
command `git duet-commit`.

This patch introduces the custom variable `magit-commit-command`. By
default the commit command is `commit`, so everything works exactly like
we're used to. But if we do `(setq magit-commit-command "duet-commit")`,
then magit will commit using git duet instead.