magit-branch-delete: fallback to delete tracking branch
When the deletion of a remote branch fails, then that's probably the
case because it no longer exists on the remote but the local tracking
branch still exists.  Users should use pruning to deal with this, but
if they don't, then this command now falls back to deleting the local
tracking branch.

Closes #2778.