Fix magit-clone (and others?) while TRAMPing
`magit-clone` would fail when called in a tramp directory. This was
specifically because the filename as passed to the the git process was
being expanded to include the magic portion of the name instead of just
the local portion.

All we need to do is test if the file is remote and get its local part
if it is.