ensure we are in correct buffer in process sentinels
In sentinels, when calling functions that have to be run in the correct
repository, always ensure that we are inside the repository by making
the buffer that is stored in the process' `command-buf' property
temporarily current.

This commit fixes two instances where we failed to do so, in
`magit-sequencer-process-sentinel' and `magit-process-finish'
(which is used by `magit-process-sentinel'.

Fixes #2805.