magit-diff-hunk-column: fix top-level case
Do not use `magit-dired-jump', that jumps to the file at point and
there might be such a file.  When there is no point at point, then
`magit-dired-jump' just jumps to `default-directory', not DIRECTORY.
This only worked because these too directories usually are the same.

Instead us `dired-jump' and append "/." to the DIRECTORY because
otherwise that function would visit the parent and place point on
DIRECTORY.