magit-section-match: fix when SECTION is nil
When SECTION is nil, then the current section is now used.  That
was already supposed to happen, but was broken.  Rename the helper
function `magit-section-match-1' to `magit-section-match-2' and
create a new `magit-section-match-1'.  Adjust related macros which
(used to) call `magit-section-match'.

Replace one use of `magit-section-when' with `magit-section-match'.
(Due to the bug fixed here, that would previously not have worked.)

Also improve the documentation of `magit-section-match' and
`magit-section-when'.