ensure we are in correct buffer in process sentinels
This builds upon the changes in [1: 4de3c98] which were added to
fix #2805.

Here we make sure that the buffer we try to use is still alive in more
cases and additionally we favor the process buffer over the command
buffer, because the latter is more likely to still be alive and all we
actually need is the appropriate value of `default-directory', which
is the same in both buffers.  Fixes #2832.

1: 4de3c9823e1214f61388eb969c5554592abac1ba
   ensure we are in correct buffer in process sentinels