magit-log-wash-rev: set author to nil for stash lines
magit-log-stash-re matches the space after the hash as the author to fit
in with the other regexps used by magit-bind-match-strings in
magit-log-wash-rev.  This, however, results in a non-nil author argument
to magit-format-log-margin.  Because magit-stashes-mode derives from
magit-reflog-mode and uses its date-only margin specifications, a
non-nil author argument pushes the date outside of the margin width in
magit-stashes-mode.

Make magit-log-wash-rev set author to nil when washing stash lines so
that magit-format-log-margin receives the expected argument of nil for
author.

See #2846.