git-commit: only override fill-column if not local yet
We used to set the buffer-local value of `fill-column' in commit
message buffers to the value of option `git-commit-fill-column',
which overrides settings in ".dir-locals.el" and is surprising.

Change `git-commit-turn-on-auto-fill' to only do so if `fill-column'
does not already have a buffer-local value and if our option's value
actually is an integer.  Also change the options default value to be
nil.  The default value of `fill-column' is 70 and that of our option
used to be 72.