magit-section-match: allow SECTION to be a type list
Support for this was dropped in [1: b2a237d], when the respective
macros where changed to pass the section instead of its type list.

However some calls to this function that still pass a type list
are likely to exist because some uses of these macros might not
have been recompiled after this function was changes.

1: b2a237d3378cb98e3493e93c1e82cc8f6d87bd78
   magit-section-match: fix when SECTION is nil