magit-popup-show-common-commands: default to nil
I recently saw some Magit screencasts and noticed that even some users
who enjoy Magit enough to create a screencast about it do not hide the
section with common commands.  Because I also think that this section
is a distraction I have decided to no longer show it by default.

This increases the odds that some users will never become aware of the
common commands.  To counter that the usage instructions displayed in
the echo area now also mention the key that can be used to show those
common commands.