re-implement support for linking to git documentation from info
Support for this was removed from `ox-texinfo+.el'.

Our implementation differs from that old implementation in that it
patches the `man' link-type instead of the `texinfo' export back-end.

As a result links have to be changed from, e.g.:

  Also see [[info:gitman#git-diff]].

to

  Also see [[man#git-diff]].

Additionally we avoid the stray space that used to appear before the
period in exported formats.  So we now actually have to write, e.g.:

  Also see [[man#git-diff]]

The new exporter now inserts the period without a proceeding space.
Unfortunately it is not possible to tell it to *not* insert a period
at all when the link does not actually end a sentence.  Fortunately
there are not many such links in our manuals (and doing this is also
perfectly in line with texinfo's restrictive formatting of links).