magit-find-file-noselect-1: check if default-directory exists
When magit-find-file and magit-find-file-other-window were introduced,
the value for the default-directory of the buffer corresponded to the
default-directory of the buffer from which the magit-find-file*
command was called.  As of e4ee279 (magit-find-{index,file}-noselect:
unify, 2016-06-22), default-directory is instead set to the file's
directory.  The new behavior has the advantage that the
default-directory is consistent for a file and doesn't depend on the
arbitrary detail of where the magit-find-file* command was called.
Also, subsequent commands that are sensitive to default-directory will
behave the same way for both a buffer visiting a revision and a buffer
visiting a file in the working tree.

It does, however, lead to problems if that directory doesn't exist in
the working tree, as many commands reasonably assume that
default-directory exists.  For example, calling dired-jump in a
revision file whose directory doesn't exist in the working tree will
signal a file-error.  Avoid these issues by using the top-level
directory if the subdirectory doesn't exist.

Re: https://github.com/magit/magit/pull/2704#r68160943