Make log margin more flexible and easier to configure
For each major-mode in whose buffers the log margin is supported add
an option `magit-*-margin', which can be customized independently of
the respective options of other modes.  `magit-log-margin-spec' used
to control some aspects of how the margin looks for all modes.

Remove the options,           adding options/variable:
`magit-duration-spec'      -> `magit--age-spec'
`magit-log-margin-spec'    ~> (all of the below:)
`magit-log-show-margin'    -> `magit-log-margin'
`magit-reflog-show-margin' -> `magit-reflog-margin'
                           -> `magit-stashes-margin'
			   -> `magit-log-select-margin'
                           -> `magit-status-margin'
`magit-refs-show-margin'   -> `magit-refs-margin'

For each of the supported major-mode it is now possible to control:
- whether the margin is shown initially,
- whether the committer name is shown,
  (though in some modes that would currently just waste space)
- and how the time is presented.

Beside showing the age of a commit, it is now also possible to show
its actual author date.

Previously the margin ended with a pseudo-fringe, but only in log
buffers, because that looked worse in other buffers.  Now no such
pseudo-fringe is used anywhere now, because it is not possible to
draw it in the area of the window past the end of the buffer.