Make selection overlay have an effect on margin when appropriate
When a sections children show information in the margin, then
the selection overlay is also used in their margin.  Until now
that did not extend to the header-line and final line of their
parent sections, causing the selection to look like this:

parent-----------|
child------------|-1 minute--
child------------|-2 minutes-
-----------------|

Now it looks like this:

parent-----------|-----------
child------------|-1 minute--
child------------|-2 minutes-
-----------------|-----------

"|" is the fringe, which may or may not look the same as "-", the
selection face.