magit-abbrev-length: handle auto-sized default length
As of Git 2.11, the default abbreviation size is no longer hardcoded
as 7 but instead scales with the repository size.

To determine the default abbreviation size in a way that works under
the old and new system, use the length of "git rev-parse --short HEAD".
Unless core.abbrev is set, this will always be at least as large as
the auto-sized value.  In the cases where HEAD needs to be lengthened
for an unambiguous id, the size will be larger than the auto-sized
value, but these situations should be rare and harmless.

Closes #2880.