magit-with-temp-index: inhibit use of the refresh cache
BODY may contain calls to git that are identical to outside calls
during the same refresh, but which are expected to return a different
value due to the sue of a different index.

Do the same in `magit-sequence-insert-sequence' around the calls to
`magit-commit-tree' and `magit-commit-worktree'.  This isn't required
at present, but if other places start calling `magit-commit-tree'
during a refresh and outside of `magit-with-temp-index', then this
would become necessary.

Re #2883.