Generalize margin support
Currently we only use the margin to display additional information
about commits, but that will likely change, not least to avoid wasting
space.  Since the format of the margin-format options changes in the
upcoming release anyway, we should take this opportunity to make the
format forward compatible and thereby avoid a similar breaking change
in the future.

The old format

  (INIT NAME DATE-STYLE)

isn't suitable, we now use

  (INIT STYLE WIDTH AUTHOR AUTHOR-WIDTH)

  (AUTHOR and AUTHOR-WIDTH replace NAME.  It is now possible to
  specify how much space should be used for the name, provided
  it is shown at all, without also showing it initially.  WIDTH
  is an addition.)

or at least that's what's being displayed in doc-strings, conceptually
the format is

  (INIT STYLE WIDTH DETAILS DETAILS-WIDTH)

The same format is used for a buffer-local variable
`magit-buffer-margin'.  Here the meaning of INIT changes to ENABLED.
This variable replaces `magit-show-margin', which said whether the
margin was enabled and several other variables, which were used to
cache values.  `magit-buffer-margin' on the other hand also caches
other information, which may be changed on a per-buffer basis.

`magit-set-buffer-margin' no longer calculates the width of the margin
itself, instead it uses WIDTH.  The option's WIDTH can be a function
whose value is then stored in `magit-buffer-margin's WIDTH.  For logs
this is always the new `magit-log-margin-width' function.  Due to the
mentioned forward compatibility this has to be exposed in the user
interface even though users should never change that for now.

There's still one hardcoded part: `magit-margin-option' returns the
option suitable for the current buffer based on a constant list.  But
that can easily be addressed when the need arises.