magit-log-refresh-popup: bind L instead of t to magit-toggle-margin
Conceptually the default command of the refresh buffer is "refresh
without saving or setting", which "g" is bound to.  However it would
not make sense to bind "L" to that action because then "L L" would be
a no-opt.

But it would still be nice to make use of the "double binding" and
the obvious candidate for that is `magit-toggle-margin'.  Previously
"t" was bound to that command, but "L t" is much less convenient than
"L L".  Also using "t" for "toggle" wasn't a good idea in the first
place; "m" for "margin" would have been much better.  Especially
when considering that the next commit will add an additional toggle
command.

I hope users will forgive me for improving this key binding, and
breaking muscle memory in the process.