magit-log-reflog-re: fix partial line match
57ce4d84 (Support showing author in all log margins, 2016-12-04)
switched the separator for magit-log-reflog-re from a space to a null
character but did not update the part used to match the trailing
partial line (which is present after "git gc" has been called).  The
failed match leads to magit-log-wash-rev signaling an error when it
encounters the partial line.

Update the regexp to use a null character separator.