magit-display-buffer-fullcolumn-most-v1: don't bury commit message
Update magit-display-buffer-fullcolumn-most-v1 to specifically handle
magit-diff-mode buffers created while committing.  This prevents the
diff buffer from filling the entire frame and hiding the commit
message.

We decided to version the names of these display functions so that we
could modify their behavior without worrying about changing something
that a user has grown used to.  However, in this case, the behavior is
clearly a defect, so let's just update v1.

Fixes #2891.