magit-log-wash-rev: correct reflog comment
The comment added in d430d856 (magit-log-wash-rev: work around `git
reflog show' bug, 2016-12-01) states that the partial trailing line
"has no basis in the data stored in the file".  However, as far as I
can tell, this partial line occurs if some reflog entries have been
expired (e.g., by 'git gc').  The reported hash on this partial line
seems to always correspond to the "old hash" field of the oldest
reflog entry.  That is, this hash is the first field of the first
entry in the corresponding .git/logs/refs/ file.

Update the comment to reflect this.

See also 81a6241e (magit-log-reflog-re: Allow for partial line,
2015-04-11).