magit-hook-custom-get: new function used to initialize hook options
Third-party hooks should not be defined as options, at least not when
they have a non-nil value, because otherwise an early `add-hook' would
not add to the hook but initialize it to just the added function.  But
users expect to be able to customize hooks, so we do it anyway.

Previously each `defcustom' form that defines a hook was preceded by a
form which made sure that the default functions got added to the hook
even when users added their own functions earlier.  Unfortunately, and
surprisingly long undiscovered, that had the side-effect of overriding
any customization made using the Custom interface.

Now we use an approach that works in both cases, obviously.

Fixes #2902.