magit-{repolist,submodule-list}-columns: fix default value
Adjust the default values so that they match the documentation and
what is expected by the Custom type of these options.  Funny enough
the code which uses these options can deal with both the old wrong
value and the value because it uses `-flatten' to turn an alist into
a plist.  Now the value isn't an alist anymore but a function whose
first and only member is an alist, but that works too and I don't
change it so that old customized values that use the wrong format
continue to function too.