Place magit-{*}-arguments in new magit-git-arguments group
Most options named `magit-TOPIC-arguments' specify the arguments
that are enabled by default in the respective `magit-TOPIC-popup'.
Previously most of these options were placed in the `magit-commands'
Custom group, which made that group rather noisy.  A few were placed
in others group, which was inconsistent and in most cases an oversight.

To avoid having to specify the group for each and every popup defined
in Magit and in third-party extensions that share the "magit-" prefix,
teach `magit-define-popup' to use `magit-git-arguments' if GROUP isn't
specified and NAME does begin with "magit-".

Also place `magit-git-global-arguments' into this group.  Remove it
from `magit', but keep it in `magit-process' as before.