magit-abbrev-length: behave reasonably in an empty repository
As of d7fcc92e (magit-abbrev-length: handle auto-sized default length,
2016-12-01), magit-abbrev-length returns 0 when called within an empty
repository or outside of a repository.

Update magit-abbrev-length to instead return 7 for these edge cases.
This value is both compatible with the default length prior to Git
v2.11 and matches the auto-sized default length that Git v2.11
calculates for a fresh repository.

Whether magit-abbrev-length should return nil or a hard-coded value
when called outside of a repository is less clear, but returning a
hard-coded value has the advantage of not requiring additional steps
to determine whether we're in a repository.  And returning 7 when
outside of a repository is consistent with the return value of
magit-abbrev-length before d7fcc92e.