Add new function `git-rebase-noop` to add a noop action at point
If you do a rebase and drop all the commits in the file, git will
do nothing, because the file is considered empty (everything is
commented). If you really do want to remove all the commits, you
have to insert a noop action which prevents the file from being
empty. It has no other purpose.