magit-rebase-insert-merge-sequence: filter out exec lines
When we collect lines from git-rebase-todo, we look for the pattern
"action hash message".  However, exec lines match this pattern if the
command consists of more than one word (or rather one word and a
trailing space), in which case the first word of the command is passed
as the hash to magit-sequence-insert-commit, leading to a type error
because magit-format-rev-summary returns nil for a non-existing
revision.

To avoid this, check that the action is not "exec" before trying to
insert the commit.  In the future, we may want to add special handling
of exec lines in order to include them in the status buffer.

Fixes #3002.