magit-rebase: fix doc-string typo
In the doc-strings of Emacs functions the names of arguments are
written in uppercase.  In Git documentation one writes the HEAD.
When we talk about the HEAD in Elisp doc-strings then we should
write `HEAD', unless we are actually referring to an argument
named HEAD.  This is consistent with how one refers to any
`variable' in a doc-string.

There are probably quite a few cases where we *don't* do that, feel
free do go looking for them.  This particular instance is being fixed
on the occasion of a user changing one head to HEAD and one `HEAD' to
HEAD, when `HEAD' would have been correct in both cases.