magit-list-refs: allow the output order to be specified
Add an option, magit-list-refs-sortby, that allows users to control
the order of refs returned by magit-list-refs and by the callers that
depend on it (magit-list-refnames, magit-list-branch-names, etc).

The most useful value for this option is probably "-creatordate".
Unlike "-committerdate" and "-taggerdate", this value works for both
commits and tags, enabling the ref to be sorted (most recent to
oldest) by the committer date if the ref is a commit or a lightweight
tag and by the tagger date if the ref is an annotated tag.

Relying on the "--sort" flag of "git for-each-ref" is somewhat
limited.  For example, a reasonable sorting pattern would be to order
by ref namespace and then by the committer or tagger date within the
namespace, but I don't think there's a way to accomplish this with
"--sort" flags.  We could allow more complex sorting patterns by
returning a list of information for each ref and passing those fields
to an arbitrary user-specified function.  However, the added expense
of this is likely not worth it.

Closes #2872, but the next few commits will be necessary to preserve
the collection order with the default completing-read and crm
completion.