git-commit-setup: improve check before inserting newline
git-commit-setup always inserts a hardcoded newline after the
beginning of a new commit message and this is very annoying when
committing with --signoff because you get by default 3 empty lines
before the "Signed-off-by:" and you have to manually remove that
hardcoded line to conform to general git log style.

This happens because `git commit` also inserts a newline (only if
--signoff is used) so in that case a redundant newline is created
(thanks tarsius for figuring this out!).

So improve the check before inserting the newline to not insert if
--signoff is used, but keep old behaviour when the flag is absent.

Signed-off-by: Ioan-Adrian Ratiu <adi@adirat.com>