Patch vc-git-conflicted-files to work around debug#21559
VC adds a hook to auto-revert for files under Git that checks whether
the current file has a merge conflict and enables smerge-mode.  This
check grabs the index lock.  When a rebase is done, Git releases the
index lock in between each cherry-pick.  This gives auto-revert a
chance to steal the lock, which causes the next operation in the
rebase to fail.

The lock is grabbed by `git status --porcelain --' run in this
function.  We change it to use `git diff-files --name-status' instead,
which does not grab the lock.

Also see the preceding commit, which imported the unpatched upstream
implementation of this function.

Re #2708 and https://debbugs.gnu.org/cgi/bugreport.cgi?bug=21559.

Thanks to Jason Merrill for the patch and Allen for pointing me toward
it.