completion: move common crm setup to a function
magit-read-range and magit-log-read-revs both use a similar crm setup,
but they differ enough that moving shared code into a common function
or macro may not be much of an improvement.  In upcoming commits,
however, we'll change the crm-related code a good amount to work
around some completion issues, and it would be cumbersome to add these
kludges in two places.