completion: work around an Emacs sorting bug
One of the functions tried by completion-all-completions,
completion-pcm--all-completions, incorrectly reverses the
collection (Emacs bug#24676).  As a result, using identity for
display-sort-function doesn't actually preserve the order of the
collection.  We can't just use nreverse instead of identity for
display-sort-function because this incorrectly reverses the collection
when the buggy path of completion-pcm--all-completions is not
followed (e.g., for partial completions).

The fix proposed in the Emacs bug report has not been applied yet.
One option is to advise completion-pcm--all-completions, reversing its
output, but that would be difficult because we don't know whether the
function will follow the buggy path.  Even if we were to duplicate the
code to check that, we would need a way to detect when the commit
lands upstream so that we don't end up reversing the correct output.
Also, advising it could interfere with users or packages that have
made a similar fix.  Instead, let's copy completion-pcm--all-completions
to a new function, and override the original during our
completing-read calls.

Along with preserving the order with the default completion system,
this solves the issue with ivy-mode and crm.  helm-mode, on the other
hand, still sorts the collection according to helm-generic-sort-fn.