Allow commit and commit when amending to be customized
The motivation for this change is to allow magit users to configure the commit commands when committing and amending a commit. (defaulting of course to "commit")

When using commit wrappers such as git-duet, it's necessary to use git duet-commit for committing, and a simple commit for amending.

By adding these custom variables, we provide the ability to use aliases and wrappers for committing and amending.