magit-shell-command-topdir: return only the read command
No longer return the directory in addition to the read command.
The callers now have to call `magit-toplevel` again.  This allows
removing the DIRECTORY argument of the callers, which should never
have made it into the function signature.  Add the new function
`magit--shell-command' to make this possible.

No longer return a list (which would now contain just the command).
The callers now have to wrap the call to this function with `list',
as is common for such reader functions.