magit-shell-command-verbose-prompt: new option
Because the commands `magit-git-command{,-topdir}' can now be used to
run arbitrary shell commands, change their prompts to not mention Git
anymore.  Use a prompt similar to that of `async-shell-command'.

By default continue to include the directory in the prompt.  That is
actually unnecessary when one of these commands is run from the popup.
Add an option `magit-shell-command-verbose-prompt' to allow users to
remove this redundancy, but don't remove it by default because some
users might invoke these commands without using the popup.