magit-stage-untracked: handle remote-less repos
If a subdirectory contains a repo that doesn't point to a remote (e.g,
a subdirectory where 'git init' was called), we pass a nil for the URL
to magit-submodule-add.  nil's are filtered out, so PATH is then taken
as the url argument by 'git submodule add', leading to a git error:

  repo URL: '.subdir/' must be absolute or begin with ./|../

Adding a submodule with a URL that points directly to a subdirectory
is probably uncommon, but Git does consider it valid.  Support this
setup by replacing a nil URL with "./<repo>".

Fixes #3047.