magit-insert-log: delimit author and date with NULs
The regexps for other log styles already use NUL as the delimiter for
the author and/or date fields.  Update magit-log-heading-re to do the
same.  For the author field, this prevents the log from choking on the
unlikely but valid case of a "[" or "]" in the author name.

Note that, unlike the previous regexp, the updated regexp assumes that
the date field exists.  But this should be safe because
magit-log-wash-rev already makes this assumption when it processes
'log' style entries without checking that the date value is non-nil.